+++ CHANGELOG.md
@@ -220,6 +220,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix `fatbuildrctl` crash when RPM spec file is not found (#165).
 - Use modern `%autopatch` macro instead of loop of `%patchN` to avoid deprecated
   syntax error during RPM packages builds (#170).
+- Fix crash of `fatbuildrctl` on missing source definition in YAML artifact
+  definition file (#171).
 - docs:
   - Add missing path parameter in REST API to retrieve artifact information.
   - Add missing optional `architectures` parameter in instances pipelines